refactor: rename 'metric' module to 'distance'
