+++ trin-core/src/portalnet/types/mod.rs
@@ -1,4 +1,4 @@
 pub mod content_key;
+pub mod distance;
 pub mod messages;
-pub mod metric;
 pub mod node;
