+++ trin-history/src/network.rs
@@ -10,8 +10,8 @@ use trin_core::{
         storage::{PortalStorage, PortalStorageConfig},
         types::{
             content_key::HistoryContentKey,
+            distance::XorMetric,
             messages::{PortalnetConfig, ProtocolId},
-            metric::XorMetric,
         },
     },
     types::validation::HeaderOracle,
